pyproject.toml
[tool.poetry]
name = "chatlib"
version = "0.7.16"
description = "A generalizable Python Chatbot Library"
authors = ["Young-Ho Kim <yghokim@younghokim.net>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.13.3"
nanoid = "^2.0.0"
python-dotenv = "^1.0.0"
pendulum = "3.0.0"
jsonlines = "^3.1.0"
yaspin = "^2.3.0"
jinja2 = "^3.1.2"
tiktoken = "^0.4.0"
transformers = "^4.37.2"
questionary = "^2.0.1"
google-generativeai = "^0.3.2"
torch = "^2.2.0"
stringcase = "^1.2.0"
requests = "^2.31.0"
anthropic = "^0.15.1"
cohere = "^4.47"
pydantic = "^2.6.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
